Refactor out an encoding item
